refactor: Move code in a distinct checkShouldStartTracking function
