+++ src/app/domain/geolocation/hooks/tracking.ts
@@ -1,16 +1,12 @@
 import { useEffect } from 'react'
 
 import { useClient } from 'cozy-client'
-import Minilog from 'cozy-minilog'
 
 import {
   isGeolocationTrackingEnabled,
-  getShouldStartTracking,
-  setGeolocationTracking
+  checkShouldStartTracking
 } from '/app/domain/geolocation/services/tracking'
 import { checkGeolocationQuota } from '/app/domain/geolocation/helpers/quota'
-
-const log = Minilog('📍 Geolocation')
 
 export const useGeolocationTracking = (): void => {
   const client = useClient()
@@ -24,13 +20,7 @@ export const useGeolocationTracking = (): void => {
       if (trackingEnabled) {
         await checkGeolocationQuota(client)
       } else {
-        const shouldStartTracking = (await getShouldStartTracking()) as boolean
-
-        if (shouldStartTracking) {
-          log.debug('Restarting geolocation tracking')
-          await setGeolocationTracking(true)
-          return
-        }
+        await checkShouldStartTracking()
       }
     }
